fix read

+++ homework03/test_water.py
@@ -6,7 +6,7 @@ class tests:
         assert water.turbid(3,4) == 12
         assert water.turbid(3,0) == 0
         assert type(water.turbid(3,4)) == int
-        assert type(water.turbid(3.1,0)) == int
+        assert type(water.turbid(3.1,0)) == float
         assert water.turbid(3.1,4) == 12.4
 
     def test_time():
